+++ .parcelrc
@@ -1,0 +1,8 @@
+{
+	"extends": "@parcel/config-default",
+  "resolvers": ["parcel-resolver-ignore", "..."],
+	"namers": ["./parcel-namer-relative-to-src.cjs"],
+  "transformers": {
+    "*.{ttf,woff,ico,webmanifest}": ["@parcel/transformer-raw"]
+  }
+}
