+++ src/icons/browserconfig.xml
