+++ src/icons/favicon-16x16.png
