+++ src/icons/favicon-32x32.png
