+++ src/icons2/manifest-icon-512.maskable.png
