+++ src/index.html
@@ -7,23 +7,23 @@
     <meta name="robots" content="index">
     <meta name="description" content="The home for everything related to in-person Treasure Hacks events">
     <title>Treasure Hacks Live</title>
-    <link rel="apple-touch-icon" sizes="180x180" href="/icons/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="/icons/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="/icons/favicon-16x16.png">
-    <link rel="manifest" href="/site.webmanifest">
-    <link rel="mask-icon" href="/icons/safari-pinned-tab.svg" color="#3c9709">
-    <link rel="shortcut icon" href="/icons/favicon.ico">
+    <link rel="apple-touch-icon" sizes="180x180" href="./icons/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="./icons/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="./icons/favicon-16x16.png">
+    <link rel="manifest" href="./site.webmanifest">
+    <link rel="mask-icon" href="./icons/safari-pinned-tab.svg" color="#3c9709">
+    <link rel="shortcut icon" href="./icons/favicon.ico">
     <meta name="msapplication-TileColor" content="#1e4a05">
-    <meta name="msapplication-config" content="/icons/browserconfig.xml">
+    <meta name="msapplication-config" content="./icons/browserconfig.xml">
     <meta name="color-scheme" content="dark">
     <meta name="theme-color" content="#2E0E36">
     <link rel="preload" href="style.css" as="style">
     <link rel="stylesheet" href="style.css">
-    <script defer src="main.js"></script>
+    <script defer type="module" src="main.js"></script>
     <!-- Linking flip: @pqina/flip -->
     <link rel="preload" href="https://unpkg.com/@pqina/flip/dist/flip.min.css" as="style">
     <link href="https://unpkg.com/@pqina/flip/dist/flip.min.css" rel="stylesheet">
-    <script defer src="https://unpkg.com/@pqina/flip/dist/flip.min.js"></script>
+    <script defer src="https://unpkg.com/@pqina/flip@1.7.7/dist/flip.min.js"></script>
 </head>
 <body>
     <header>
